backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
PyPDF2
python-docx
python-multipart